Refactor to create abstract project init functions, making it backend agnostic
