+++ src/usethis/_core/ci.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from usethis._console import box_print, info_print
-from usethis._integrations.backend.uv.init import ensure_pyproject_toml
+from usethis._init import ensure_pyproject_toml
 from usethis._integrations.ci.bitbucket.config import (
     add_bitbucket_pipeline_config,
     remove_bitbucket_pipeline_config,
