+++ src/usethis/_core/show.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from usethis._config import usethis_config
 from usethis._console import plain_print
-from usethis._integrations.backend.uv.init import ensure_pyproject_toml
+from usethis._init import ensure_pyproject_toml
 from usethis._integrations.project.name import get_project_name
 from usethis._integrations.sonarqube.config import get_sonar_project_properties
