+++ src/usethis/_core/status.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from usethis._console import tick_print
-from usethis._integrations.backend.uv.init import ensure_pyproject_toml
+from usethis._init import ensure_pyproject_toml
 from usethis._integrations.file.pyproject_toml.io_ import PyprojectTOMLManager
 from usethis._types.status import DevelopmentStatusEnum
