+++ tests/usethis/_core/test_docstyle.py
@@ -5,7 +5,7 @@ import pytest
 from usethis._config_file import files_manager
 from usethis._core.docstyle import use_docstyle
 from usethis._core.tool import use_ruff
-from usethis._integrations.backend.uv.init import ensure_pyproject_toml
+from usethis._init import ensure_pyproject_toml
 from usethis._test import change_cwd
 from usethis._types.docstyle import DocStyleEnum
